Update README.md
I've changed the link to the python script. Now if you right click one the link, you can download the script directly. The previous link also lead to the script, but if you download it, it turns out as a website.

I also fix one link to the database.

Zewei