Added sort option; fixed newline bug
The sort option makes sorting optional.
Somtimes a table might have more columns than just the otu id, count columns, and taxonmy column.
When there is an extra column that is not numeric, you get an error caused by trying to sum it for the sort (see line 305).

I also found a bug where there were extra newlines carried over from the input files.
This was making the output invalid.
I solved this by removing all newline when files are read and putting them back in when outputs are written.